feat(uploader): init unified uploader for transcoder
The problem that we are trying to solve here is that in the monolith mode, getting PUT url, changing the db status are trivial by calling the function. In the worker mode, we want the worker to be stateless for simplicity and security. Worker needs to call the main api for any mutation.